refactor(tools): replace skill management tools with skill_preview
Replace create_skill/update_skill/fork_skill (web-only, DB-coupled) with
a single skill_preview tool that generates structured skill content as an
inline collapsible card with per-file copy buttons.

Changes:
- Add skill_preview tool (web/channel only) — pure content generator,
  returns structured JSON rendered by frontend as collapsible card
- Add SkillPreviewCard.tsx — three-level UI: skill → files → content,
  each file has a Copy button (navigator.clipboard)
- Add system prompt guidance: web/channel uses skill_preview tool,
  cli mode uses fenced code blocks for copyable output
- Delete create-skill.ts, update-skill.ts, fork-skill.ts
- Delete SkillPanel.tsx (sidebar), SkillCard.tsx (entry card)
- Delete platform skills: create-skill/SKILL.md, update-skill/SKILL.md
- Update docs: tools.md, skills.md, architecture.md
